+++ adapters/gnu.pl
@@ -1,7 +1,7 @@
 %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
 %
 %  Adapter file for GNU Prolog 1.4.2 (and later versions)
-%  Last updated on May 5, 2017
+%  Last updated on August 26, 2017
 %
 %  This file is part of Logtalk <http://logtalk.org/>  
 %  Copyright 1998-2017 Paulo Moura <pmoura@logtalk.org>
@@ -606,7 +606,6 @@ thread_join(_, _) :- fail.
 with_mutex(_, _) :- fail.
 message_queue_create(_) :- fail.
 message_queue_destroy(_) :- fail.
-:- dynamic('$lgt_engine_term_queue_'/2).