+++ adapters/lean.pl
@@ -1,7 +1,7 @@
 %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
 %
 %  Adapter file for Lean Prolog 4.5.7 and later versions
-%  Last updated on May 5, 2017
+%  Last updated on August 26, 2017
 %
 %  This file is part of Logtalk <http://logtalk.org/>  
 %  Copyright 1998-2017 Paulo Moura <pmoura@logtalk.org>
@@ -687,8 +687,6 @@ message_queue_create(_) :-
 
 message_queue_destroy(_) :-
 	fail.
-
-:- dynamic('$lgt_engine_term_queue_'/2).